integration: extract timestamps.rs
Tests `tests_timestamp()` and `test_timestamp_generator()` are extracted
from `session/session.rs` into `statement/timestamps.rs`
