+++ scylla/tests/integration/statements/mod.rs
@@ -4,3 +4,4 @@ mod execution_profiles;
 mod silent_prepare_query;
 mod skip_metadata_optimization;
 mod statement;
+mod timestamps;
